Add NHWC conversion as a graph transformer pass
Modelled off of the NCHWc conversion. The advantage of doing this via onnxruntime is that we can run models directly produced by e.g. pytorch and we'll automatically do the conversion in runtime to reap the benefits. Of course, this is a lot slower (especially because c++ libraries for tensor manipulation are terrible compared to numpy), but onnxruntime can save the transformed model to disk.