Convert Gemm with TransB into MatMul + Add
This allows all ops of Resnet50 to be run on Gemmini
However, converting quantizing the last MatMul hurts accuracy
quite a bit. Intuitively I suspect this is because the weight matrix
for the matmul (gpu_0/pred_w_0) has elements in a very small range,
and these are esentially what directly control the predictions.