Fix bug with shape inference and variable dims
ORT uses the shape inference data to determine how to do memory
allocation. Mask-RCNN allows the height and width of the image to be
variable, but we weren't checking for that when we do shape inference.
Thus the height and width we obtained during shape inference were set to
the default value of 1, and this meant that tensors which weren't
actually the same dimensions had the same inferred dimension, resulting
in errors when the memory allocator tried to reuse a buffer that it
couldn't